Resolve the findbugs annotations vs. jsr305 war
dependency plugin 2.10 will hopefully allow exclusion of dependencies
from checking, so basepom now will tolerate bot jars on the
classpath. This is not ideal but the best what can be done without
continously force everyone into exclusions all the time.

This pull request can be merged as soon as it passes. :-)